#18 Structure update
- The bid module structure has been updated
