+++ bid-service/src/main/java/van/karm/bid/application/service/BidService.java
@@ -1,0 +1,8 @@
+package van.karm.bid.application.service;
+
+import org.springframework.security.oauth2.jwt.Jwt;
+import van.karm.bid.presentation.dto.request.AddBid;
+
+public interface BidService {
+    void addBid(Jwt jwt, AddBid bid);
+}
