+++ bid-service/src/main/java/van/karm/bid/application/validator/AuctionValidator.java
@@ -1,0 +1,7 @@
+package van.karm.bid.application.validator;
+
+import van.karm.bid.presentation.dto.request.AddBid;
+
+public interface AuctionValidator {
+    boolean isValid(String userId, AddBid bid);
+}
