--- bid-service/src/main/java/van/karm/bid/config/SecurityConfig.java
@@ -1,53 +1,0 @@
-package van.karm.bid.config;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.Customizer;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.web.SecurityFilterChain;
-import van.karm.bid.config.exception.RestAccessDeniedHandler;
-import van.karm.bid.config.exception.RestAuthenticationEntryPoint;
-
-@RequiredArgsConstructor
-@Configuration
-@EnableWebSecurity
-public class SecurityConfig {
-    private final RestAccessDeniedHandler accessDeniedHandler;
-    private final RestAuthenticationEntryPoint authenticationEntryPoint;
-    private static final String[] ACTUATOR_WHITELIST = {
-            "/actuator",
-            "/actuator/**"
-    };
-    private static final String[] SWAGGER_WHITELIST = {
-            "/v3/api-docs/**",
-            "/swagger-ui.html",
-            "/swagger-ui/**",
-            "/webjars/**",
-            "/swagger-resources/**",
-            "/swagger-resources/configuration/ui",
-            "/swagger-resources/configuration/security",
-            "/openapi.yaml"
-    };
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        return http
-                .csrf(AbstractHttpConfigurer::disable)
-                .authorizeHttpRequests(auth -> auth
-                        .requestMatchers(ACTUATOR_WHITELIST).permitAll()
-                        .requestMatchers(SWAGGER_WHITELIST).permitAll()
-                        .anyRequest().authenticated()
-                )
-                .exceptionHandling(exception -> exception
-                        .accessDeniedHandler(accessDeniedHandler)
-                )
-                .oauth2ResourceServer(oauth2 -> oauth2
-                        .jwt(Customizer.withDefaults())
-                        .authenticationEntryPoint(authenticationEntryPoint)
-                )
-                .build();
-    }
-}
