--- bid-service/src/main/java/van/karm/bid/config/exception/RestAccessDeniedHandler.java
@@ -1,30 +1,0 @@
-package van.karm.bid.config.exception;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-import java.time.Instant;
-
-@Component
-public class RestAccessDeniedHandler implements AccessDeniedHandler {
-    @Override
-    public void handle(HttpServletRequest request,
-                       HttpServletResponse response,
-                       AccessDeniedException accessDeniedException) throws IOException {
-        response.setContentType("application/json");
-        response.setStatus(HttpServletResponse.SC_FORBIDDEN);
-        response.getWriter().write("""
-            {
-              "status":403,
-              "error":"Forbidden",
-              "message":"%s",
-              "path":"%s",
-              "timestamp":"%s"
-            }
-            """.formatted(accessDeniedException.getMessage(), request.getRequestURI(), Instant.now()));
-    }
-}
