--- bid-service/src/main/java/van/karm/bid/config/exception/RestAuthenticationEntryPoint.java
@@ -1,30 +1,0 @@
-package van.karm.bid.config.exception;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-import java.time.Instant;
-
-@Component
-public class RestAuthenticationEntryPoint implements AuthenticationEntryPoint {
-    @Override
-    public void commence(HttpServletRequest request,
-                         HttpServletResponse response,
-                         AuthenticationException authException) throws IOException {
-        response.setContentType("application/json");
-        response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-        response.getWriter().write("""
-            {
-              "status":401,
-              "error":"Unauthorized",
-              "message":"%s",
-              "path":"%s",
-              "timestamp":"%s"
-            }
-            """.formatted(authException.getMessage(), request.getRequestURI(), Instant.now()));
-    }
-}
