--- bid-service/src/main/java/van/karm/bid/controller/exception/ApiError.java
@@ -1,11 +1,0 @@
-package van.karm.bid.controller.exception;
-
-import java.time.Instant;
-
-public record ApiError(
-        int status,
-        String error,
-        String message,
-        String path,
-        Instant timestamp
-) {}
