+++ bid-service/src/main/java/van/karm/bid/domain/exception/AccessDeniedException.java
@@ -1,0 +1,7 @@
+package van.karm.bid.domain.exception;
+
+public class AccessDeniedException extends RuntimeException {
+    public AccessDeniedException(String message) {
+        super(message);
+    }
+}
