+++ bid-service/src/main/java/van/karm/bid/domain/exception/InvalidArgumentException.java
@@ -1,0 +1,7 @@
+package van.karm.bid.domain.exception;
+
+public class InvalidArgumentException extends RuntimeException {
+    public InvalidArgumentException(String message) {
+        super(message);
+    }
+}
