+++ bid-service/src/main/java/van/karm/bid/domain/exception/InvalidTokenException.java
@@ -1,0 +1,7 @@
+package van.karm.bid.domain.exception;
+
+public class InvalidTokenException extends RuntimeException {
+    public InvalidTokenException(String message) {
+        super(message);
+    }
+}
