+++ bid-service/src/main/java/van/karm/bid/domain/exception/ServiceUnavailableException.java
@@ -1,0 +1,7 @@
+package van.karm.bid.domain.exception;
+
+public class ServiceUnavailableException extends RuntimeException {
+    public ServiceUnavailableException(String message) {
+        super(message);
+    }
+}
