+++ bid-service/src/main/java/van/karm/bid/domain/exception/UnauthenticatedException.java
@@ -1,0 +1,7 @@
+package van.karm.bid.domain.exception;
+
+public class UnauthenticatedException extends RuntimeException {
+    public UnauthenticatedException(String message) {
+        super(message);
+    }
+}
