--- bid-service/src/main/java/van/karm/bid/enums/Auction/AuctionStatus.java
@@ -1,7 +1,0 @@
-package van.karm.bid.enums.Auction;
-
-public enum AuctionStatus {
-    ACTIVE,
-    INACTIVE,
-    FINISHED
-}
