--- bid-service/src/main/java/van/karm/bid/exception/AccessDeniedException.java
@@ -1,7 +1,0 @@
-package van.karm.bid.exception;
-
-public class AccessDeniedException extends RuntimeException {
-    public AccessDeniedException(String message) {
-        super(message);
-    }
-}
