--- bid-service/src/main/java/van/karm/bid/exception/InvalidArgumentException.java
@@ -1,7 +1,0 @@
-package van.karm.bid.exception;
-
-public class InvalidArgumentException extends RuntimeException {
-    public InvalidArgumentException(String message) {
-        super(message);
-    }
-}
