--- bid-service/src/main/java/van/karm/bid/exception/InvalidTokenException.java
@@ -1,7 +1,0 @@
-package van.karm.bid.exception;
-
-public class InvalidTokenException extends RuntimeException {
-    public InvalidTokenException(String message) {
-        super(message);
-    }
-}
