--- bid-service/src/main/java/van/karm/bid/exception/ServiceUnavailableException.java
@@ -1,7 +1,0 @@
-package van.karm.bid.exception;
-
-public class ServiceUnavailableException extends RuntimeException {
-    public ServiceUnavailableException(String message) {
-        super(message);
-    }
-}
