--- bid-service/src/main/java/van/karm/bid/exception/UnauthenticatedException.java
@@ -1,7 +1,0 @@
-package van.karm.bid.exception;
-
-public class UnauthenticatedException extends RuntimeException {
-    public UnauthenticatedException(String message) {
-        super(message);
-    }
-}
