+++ bid-service/src/main/java/van/karm/bid/presentation/dto/response/ApiError.java
@@ -1,0 +1,11 @@
+package van.karm.bid.presentation.dto.response;
+
+import java.time.Instant;
+
+public record ApiError(
+        int status,
+        String error,
+        String message,
+        String path,
+        Instant timestamp
+) {}
