--- bid-service/src/main/java/van/karm/bid/repo/BidRepo.java
@@ -1,9 +1,0 @@
-package van.karm.bid.repo;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import van.karm.bid.model.Bid;
-
-import java.util.UUID;
-
-public interface BidRepo extends JpaRepository<Bid, UUID> {
-}
