--- bid-service/src/main/java/van/karm/bid/service/BidService.java
@@ -1,8 +1,0 @@
-package van.karm.bid.service;
-
-import org.springframework.security.oauth2.jwt.Jwt;
-import van.karm.bid.dto.request.AddBid;
-
-public interface BidService {
-    void addBid(Jwt jwt, AddBid bid);
-}
